+++ 1.20/Fabric/src/main/java/fuzs/enchantinginfuser/core/FabricAbstractions.java
@@ -14,5 +14,4 @@ public final class FabricAbstractions implements CommonAbstractions {
     public boolean isAllowedOnBooks(Enchantment enchantment) {
         return true;
     }
-
 }