+++ 1.20/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.2-1.20.1] - 2023-10-04
+### Changed
+- Re-enabled integration for Apotheosis on Forge
+### Fixed
+- Fixed disenchanting sometimes giving back more experience than was required for the enchantment in the first place
+- Fixed being able to change enchantments on some unenchantable items
+
 ## [v8.0.1-1.20.1] - 2023-07-19
 ### Changed
 - The advanced infuser can now only repair tools and armor items to prevent exploits with items from other mods that have durability